# Update JsonWriter performance tests and add more unit tests.