--- a/src/System.Text.JsonLab/System/Text/Json/JsonWriterUtf16.cs
+++ b/src/System.Text.JsonLab/System/Text/Json/JsonWriterUtf16.cs
@@ -6,12 +6,12 @@ using System.Buffers.Text;
 using System.Diagnostics;
 using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
-using System.Text.Formatting;
 
 namespace System.Text.JsonLab
 {
     public struct JsonWriterUtf16
     {
+        private static readonly char[] s_newLine = Environment.NewLine.ToCharArray();
         readonly bool _prettyPrint;
         readonly IBufferWriter<byte> _bufferWriter;
 
@@ -93,18 +93,7 @@ namespace System.Text.JsonLab
                 charBuffer[idx++] = (char)JsonConstants.ListSeperator;
 
             if (_prettyPrint)
-            {
-                charBuffer[idx++] = (char)JsonConstants.CarriageReturn;
-                charBuffer[idx++] = (char)JsonConstants.LineFeed;
-
-                int indent = _indent;
-
-                while (indent-- >= 0)
-                {
-                    charBuffer[idx++] = (char)JsonConstants.Space;
-                    charBuffer[idx++] = (char)JsonConstants.Space;
-                }
-            }
+                AddNewLineAndIndentation(charBuffer, ref idx);
 
             charBuffer[idx++] = (char)JsonConstants.Quote;
 
@@ -142,18 +131,7 @@ namespace System.Text.JsonLab
             int idx = 0;
 
             if (_prettyPrint)
-            {
-                charBuffer[idx++] = (char)JsonConstants.CarriageReturn;
-                charBuffer[idx++] = (char)JsonConstants.LineFeed;
-
-                int indent = _indent;
-
-                while (indent-- >= 0)
-                {
-                    charBuffer[idx++] = (char)JsonConstants.Space;
-                    charBuffer[idx++] = (char)JsonConstants.Space;
-                }
-            }
+                AddNewLineAndIndentation(charBuffer, ref idx);
 
             charBuffer[idx] = (char)token;
             _bufferWriter.Advance(bytesNeeded);
@@ -221,18 +199,7 @@ namespace System.Text.JsonLab
                 charBuffer[idx++] = (char)JsonConstants.ListSeperator;
 
             if (_prettyPrint)
-            {
-                charBuffer[idx++] = (char)JsonConstants.CarriageReturn;
-                charBuffer[idx++] = (char)JsonConstants.LineFeed;
-
-                int indent = _indent;
-
-                while (indent-- >= 0)
-                {
-                    charBuffer[idx++] = (char)JsonConstants.Space;
-                    charBuffer[idx++] = (char)JsonConstants.Space;
-                }
-            }
+                AddNewLineAndIndentation(charBuffer, ref idx);
 
             charBuffer[idx++] = (char)JsonConstants.Quote;
 
@@ -370,18 +337,7 @@ namespace System.Text.JsonLab
                 charBuffer[idx++] = (char)JsonConstants.ListSeperator;
 
             if (_prettyPrint)
-            {
-                charBuffer[idx++] = (char)JsonConstants.CarriageReturn;
-                charBuffer[idx++] = (char)JsonConstants.LineFeed;
-
-                int indent = _indent;
-
-                while (indent-- >= 0)
-                {
-                    charBuffer[idx++] = (char)JsonConstants.Space;
-                    charBuffer[idx++] = (char)JsonConstants.Space;
-                }
-            }
+                AddNewLineAndIndentation(charBuffer, ref idx);
 
             charBuffer[idx++] = (char)JsonConstants.Quote;
 
@@ -424,18 +380,7 @@ namespace System.Text.JsonLab
 
             _firstItem = false;
             if (_prettyPrint)
-            {
-                charBuffer[idx++] = (char)JsonConstants.CarriageReturn;
-                charBuffer[idx++] = (char)JsonConstants.LineFeed;
-
-                int indent = _indent;
-
-                while (indent-- >= 0)
-                {
-                    charBuffer[idx++] = (char)JsonConstants.Space;
-                    charBuffer[idx++] = (char)JsonConstants.Space;
-                }
-            }
+                AddNewLineAndIndentation(charBuffer, ref idx);
 
             if (insertNegationSign)
                 charBuffer[idx++] = '-';
@@ -615,18 +560,7 @@ namespace System.Text.JsonLab
                 charBuffer[idx++] = (char)JsonConstants.ListSeperator;
 
             if (_prettyPrint)
-            {
-                charBuffer[idx++] = (char)JsonConstants.CarriageReturn;
-                charBuffer[idx++] = (char)JsonConstants.LineFeed;
-
-                int indent = _indent;
-
-                while (indent-- >= 0)
-                {
-                    charBuffer[idx++] = (char)JsonConstants.Space;
-                    charBuffer[idx++] = (char)JsonConstants.Space;
-                }
-            }
+                AddNewLineAndIndentation(charBuffer, ref idx);
 
             charBuffer[idx++] = (char)JsonConstants.Quote;
 
@@ -793,7 +727,7 @@ namespace System.Text.JsonLab
 
             if (_prettyPrint)
             {
-                int bytesNeededForPrettyPrint = 2;  // For the new line, \r\n
+                int bytesNeededForPrettyPrint = s_newLine.Length;  // For the new line, \r\n or \n
                 bytesNeededForPrettyPrint += (_indent + 1) * 2;
                 bytesNeeded += numBytes * bytesNeededForPrettyPrint;
             }
@@ -809,7 +743,7 @@ namespace System.Text.JsonLab
 
             if (_prettyPrint)
             {
-                int bytesNeededForPrettyPrint = 2;  // For the new line, \r\n
+                int bytesNeededForPrettyPrint = s_newLine.Length;  // For the new line, \r\n or \n
                 bytesNeededForPrettyPrint += (_indent + 1) * 2;
                 bytesNeeded += numBytes * bytesNeededForPrettyPrint;
             }
@@ -826,7 +760,7 @@ namespace System.Text.JsonLab
 
             if (_prettyPrint)
             {
-                int bytesNeededForPrettyPrint = 2;    // For the new line, \r\n
+                int bytesNeededForPrettyPrint = s_newLine.Length;    // For the new line, \r\n or \n
                 bytesNeededForPrettyPrint += (_indent + 1) * 2;
                 bytesNeeded += numBytes * bytesNeededForPrettyPrint;
             }
@@ -896,6 +830,28 @@ namespace System.Text.JsonLab
             bytesNeeded += MemoryMarshal.AsBytes(valueSpan).Length;
 
             return bytesNeeded;
+        }
+
+        private void AddNewLineAndIndentation(Span<char> buffer, ref int idx)
+        {
+            // \r\n versus \n, depending on OS
+            if (s_newLine.Length == 2)
+            {
+                buffer[idx++] = s_newLine[0];   // \r
+                buffer[idx++] = s_newLine[1];   // \n
+            }
+            else
+            {
+                buffer[idx++] = s_newLine[0];   // \n
+            }
+
+            int indent = _indent;
+
+            while (indent-- >= 0)
+            {
+                buffer[idx++] = (char)JsonConstants.Space;
+                buffer[idx++] = (char)JsonConstants.Space;
+            }
         }
     }
 }