--- a/src/System.Text.JsonLab/System/Text/Json/JsonWriterUtf16.cs
+++ b/src/System.Text.JsonLab/System/Text/Json/JsonWriterUtf16.cs
@@ -780,7 +780,7 @@ namespace System.Text.JsonLab
 
             if (_prettyPrint)
             {
-                int bytesNeededForPrettyPrint = 3;    // For the new line, \r\n, and the space after the colon
+                int bytesNeededForPrettyPrint = s_newLine.Length + 1;    // For the new line, \r\n or \n, and the space after the colon
                 bytesNeededForPrettyPrint += (_indent + 1) * 2;
                 bytesNeeded += numBytes * bytesNeededForPrettyPrint;
             }
@@ -800,7 +800,7 @@ namespace System.Text.JsonLab
 
             if (_prettyPrint)
             {
-                int bytesNeededForPrettyPrint = 3;    // For the new line, \r\n,  and the space after the colon
+                int bytesNeededForPrettyPrint = s_newLine.Length + 1;    // For the new line, \r\n or \n,  and the space after the colon
                 bytesNeededForPrettyPrint += (_indent + 1) * 2;
                 bytesNeeded += numBytes * bytesNeededForPrettyPrint;
             }
@@ -820,7 +820,7 @@ namespace System.Text.JsonLab
 
             if (_prettyPrint)
             {
-                int bytesNeededForPrettyPrint = 3;    // For the new line, \r\n,  and the space after the colon
+                int bytesNeededForPrettyPrint = s_newLine.Length + 1;    // For the new line, \r\n or \n,  and the space after the colon
                 bytesNeededForPrettyPrint += (_indent + 1) * 2;
                 bytesNeeded += numBytes * bytesNeededForPrettyPrint;
             }